+++ polaris-migrator/src/migrations/react-replace-text-components/react-replace-text-components.ts
@@ -3,11 +3,12 @@ import type {API, FileInfo, Options} from 'jscodeshift';
 import {hasImportDeclaration} from '../../utilities/imports';
 
 import {replaceDisplayText} from './steps/replace-display-text';
-import {replaceOther} from './steps/replace-other';
+import {replaceOther, components} from './steps/replace-other';
 import {replaceTextStyle} from './steps/replace-text-style';
 
 export interface MigrationOptions extends Options {
   relative: boolean;
+  componentName?: keyof typeof components | 'DisplayText' | 'TextStyle';
 }
 
 export default function reactReplaceTextComponents(
@@ -24,9 +25,19 @@ export default function reactReplaceTextComponents(
     return file.source;
   }
 
-  replaceDisplayText(j, source, options);
-  replaceOther(j, source, options);
-  replaceTextStyle(j, source, options);
+  if (options.componentName) {
+    if (options.componentName === 'DisplayText') {
+      replaceDisplayText(j, source, options);
+    } else if (options.componentName === 'TextStyle') {
+      replaceTextStyle(j, source, options);
+    } else {
+      replaceOther(j, source, options);
+    }
+  } else {
+    replaceDisplayText(j, source, options);
+    replaceOther(j, source, options);
+    replaceTextStyle(j, source, options);
+  }
 
   return source.toSource();
 }