+++ polaris-react/src/components/Modal/Modal.stories.tsx
@@ -1,10 +1,12 @@
 import React, {useCallback, useRef, useState} from 'react';
 import type {ComponentMeta} from '@storybook/react';
 import {
+  Banner,
   Button,
   Checkbox,
   ChoiceList,
   DropZone,
+  FormLayout,
   Modal,
   Stack,
   Text,
@@ -478,5 +480,102 @@ export function WithoutAnActivatorProp() {
         </Modal.Section>
       </Modal>
     </div>
+  );
+}
+
+export function WithLongContent() {
+  return (
+    <Modal
+      title="Long form modal"
+      open
+      onClose={() => {}}
+      primaryAction={{content: 'Save'}}
+    >
+      <Modal.Section>
+        <Banner title="Payment details" />
+      </Modal.Section>
+      <Modal.Section>
+        <FormLayout>
+          <FormLayout.Group>
+            <TextField label="Payment method type" autoComplete="off" />
+            <TextField label="Card number" autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="Expires" autoComplete="off" />
+            <TextField label="CVV" autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="Country/region" autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="First name" autoComplete="off" />
+            <TextField label="Last name" autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="Address" autoComplete="off" />
+            <TextField label="Apartment, suite, etc." autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="City" autoComplete="off" />
+            <TextField label="Province" autoComplete="off" />
+            <TextField label="Postal code" autoComplete="off" />
+          </FormLayout.Group>
+        </FormLayout>
+      </Modal.Section>
+    </Modal>
+  );
+}
+
+export function WithLongContentNoScroll() {
+  return (
+    <Modal
+      title="Long form modal"
+      open
+      onClose={() => {}}
+      noScroll
+      primaryAction={{content: 'Save'}}
+    >
+      <Modal.Section>
+        <Banner title="Payment details" />
+      </Modal.Section>
+      <Modal.Section>
+        <FormLayout>
+          <FormLayout.Group>
+            <TextField label="Payment method type" autoComplete="off" />
+            <TextField label="Card number" autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="Expires" autoComplete="off" />
+            <TextField label="CVV" autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="Country/region" autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="First name" autoComplete="off" />
+            <TextField label="Last name" autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="Address" autoComplete="off" />
+            <TextField label="Apartment, suite, etc." autoComplete="off" />
+          </FormLayout.Group>
+
+          <FormLayout.Group>
+            <TextField label="City" autoComplete="off" />
+            <TextField label="Province" autoComplete="off" />
+            <TextField label="Postal code" autoComplete="off" />
+          </FormLayout.Group>
+        </FormLayout>
+      </Modal.Section>
+    </Modal>
   );
 }