+++ polaris-react/src/components/Modal/components/Footer/Footer.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import type {ComplexAction} from '../../../../types';
 import {buttonsFrom} from '../../../Button';
-import {ButtonGroup} from '../../../ButtonGroup';
 import {Box} from '../../../Box';
 import {Inline} from '../../../Inline';
 
@@ -26,25 +25,25 @@ export function Footer({
     (secondaryActions && buttonsFrom(secondaryActions)) || null;
   const actions =
     primaryActionButton || secondaryActionButtons ? (
-      <ButtonGroup>
+      <Inline gap="2">
         {secondaryActionButtons}
         {primaryActionButton}
-      </ButtonGroup>
+      </Inline>
     ) : null;
 
   return (
-    <Box
-      borderBlockStart="divider"
-      minHeight="var(--p-space-16)"
-      padding="4"
-      width="100%"
-    >
-      <Inline align="space-between">
-        <Inline blockAlign="center">{children}</Inline>
-        <Inline align="end" blockAlign="center">
+    <Inline blockAlign="center">
+      <Box
+        borderBlockStart="divider"
+        minHeight="var(--p-space-16)"
+        padding="4"
+        width="100%"
+      >
+        <Inline blockAlign="center" align="space-between">
+          <Box>{children}</Box>
           {actions}
         </Inline>
-      </Inline>
-    </Box>
+      </Box>
+    </Inline>
   );
 }